Autochanger for read + 64 bit addrs + Session key, see kes14Sep02
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@150 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/stored/dev.c
@@ -404,7 +404,7 @@ int update_pos_dev(DEVICE *dev)
 #ifdef xxxx
    struct mtget mt_stat;
 #endif
-   int32_t pos;
+   int64_t pos;
    int stat = 0;
 
    if (dev->fd < 0) {
